# Faster setZero (Fixes #13692)