Motivation:

Unsafe's buffer setZero always piggyback to the JDK's Unsafe::setMemory,
which is not optimized for small writes, given that's not an intrinsic candidate
and requires handling thread-state transitions.

Modifications:

Handle small writes with a specialized code path which won't rely on Unsafe::setMemory

Result:

Faster setZero(...) and writeZero(...)