@franz1981 2023-11-09
@normanmaurer I'm adding 2 commits that @imangry  could play and try with his own micro (fixed after my comments!)