@franz1981 2023-11-09
@normanmaurer I've changed my mind: let's see first if this PR already deliver what's promised and we can see if the check for  when apply this optimization is still 64 or just 8; and eventually adding int/short handling OR just the int handling too