@franz1981 2023-11-10
@normanmaurer and @imangry these are the results for the latest commit

```
SINGLE BYTES, DIRECT BUFFERS:

Benchmark                         (bytes)  (checkAccessible)  (checkBounds)  (direct)  (pooled)  (startOffset)  Mode  Cnt   Score   Error  Units
ByteBufZeroingBenchmark.setBytes        1              false          false      true      true              0  avgt   10   1.485 ± 0.002  ns/op
ByteBufZeroingBenchmark.setBytes        2              false          false      true      true              0  avgt   10   1.784 ± 0.008  ns/op
ByteBufZeroingBenchmark.setBytes        3              false          false      true      true              0  avgt   10   2.763 ± 0.070  ns/op
ByteBufZeroingBenchmark.setBytes        4              false          false      true      true              0  avgt   10   3.216 ± 0.007  ns/op
ByteBufZeroingBenchmark.setBytes        7              false          false      true      true              0  avgt   10   4.175 ± 0.045  ns/op
ByteBufZeroingBenchmark.setBytes        8              false          false      true      true              0  avgt   10   3.883 ± 0.041  ns/op
ByteBufZeroingBenchmark.setBytes       15              false          false      true      true              0  avgt   10   4.811 ± 0.095  ns/op
ByteBufZeroingBenchmark.setBytes       64              false          false      true      true              0  avgt   10   7.269 ± 0.021  ns/op
ByteBufZeroingBenchmark.setBytes       65              false          false      true      true              0  avgt   10   7.353 ± 0.135  ns/op
ByteBufZeroingBenchmark.setBytes     1024              false          false      true      true              0  avgt   10  28.634 ± 0.095  ns/op

BEFORE, DIRECT BUFFERS:

Benchmark                         (bytes)  (checkAccessible)  (checkBounds)  (direct)  (pooled)  (startOffset)  Mode  Cnt   Score   Error  Units
ByteBufZeroingBenchmark.setZero         1              false          false      true      true              0  avgt   10   8.023 ± 0.040  ns/op
ByteBufZeroingBenchmark.setZero         2              false          false      true      true              0  avgt   10   7.583 ± 0.007  ns/op
ByteBufZeroingBenchmark.setZero         3              false          false      true      true              0  avgt   10   8.027 ± 0.018  ns/op
ByteBufZeroingBenchmark.setZero         4              false          false      true      true              0  avgt   10   7.630 ± 0.078  ns/op
ByteBufZeroingBenchmark.setZero         7              false          false      true      true              0  avgt   10  10.780 ± 3.705  ns/op
ByteBufZeroingBenchmark.setZero         8              false          false      true      true              0  avgt   10   7.426 ± 0.056  ns/op
ByteBufZeroingBenchmark.setZero        15              false          false      true      true              0  avgt   10   8.292 ± 0.268  ns/op
ByteBufZeroingBenchmark.setZero        64              false          false      true      true              0  avgt   10   8.835 ± 0.605  ns/op
ByteBufZeroingBenchmark.setZero        65              false          false      true      true              0  avgt   10   8.869 ± 0.486  ns/op
ByteBufZeroingBenchmark.setZero      1024              false          false      true      true              0  avgt   10  22.767 ± 0.182  ns/op


AFTER, DIRECT BUFFERS:

Benchmark                        (bytes)  (checkAccessible)  (checkBounds)  (direct)  (pooled)  (startOffset)  Mode  Cnt   Score   Error  Units
ByteBufZeroingBenchmark.setZero        1              false          false      true      true              0  avgt   10   1.795 ± 0.023  ns/op
ByteBufZeroingBenchmark.setZero        2              false          false      true      true              0  avgt   10   2.013 ± 0.017  ns/op
ByteBufZeroingBenchmark.setZero        3              false          false      true      true              0  avgt   10   3.264 ± 0.047  ns/op
ByteBufZeroingBenchmark.setZero        4              false          false      true      true              0  avgt   10   3.345 ± 0.007  ns/op
ByteBufZeroingBenchmark.setZero        7              false          false      true      true              0  avgt   10   3.643 ± 0.027  ns/op
ByteBufZeroingBenchmark.setZero        8              false          false      true      true              0  avgt   10   1.821 ± 0.016  ns/op
ByteBufZeroingBenchmark.setZero       15              false          false      true      true              0  avgt   10   4.907 ± 0.007  ns/op
ByteBufZeroingBenchmark.setZero       64              false          false      true      true              0  avgt   10   3.755 ± 0.025  ns/op
ByteBufZeroingBenchmark.setZero       65              false          false      true      true              0  avgt   10   9.375 ± 1.166  ns/op
ByteBufZeroingBenchmark.setZero     1024              false          false      true      true              0  avgt   10  22.742 ± 0.014  ns/op
```
The improvement seems dramatic (but need to take in account that the accessed buffer's memory is always in L1 cache, and hot, which is  not what always happen in real world!), for smaller write size as expected, which seems to be due to the presence of the `Unsafe::setMemory` AND `memset` calls for small writes (of 1 byte!) which basically become a benchmark of 2 not-inlined calls vs a branch + put byte (that's why the diff is ~5 times more!).
Reason for that is that `Unsafe::setMemory` isn't an intrinsic nor OpenJDK users are directly exposed on it (it's used on the DirectBuffer's constructor, which is not an hot spot!).

The test has run on `# VM version: JDK 17.0.7, Java HotSpot(TM) 64-Bit Server VM, 17.0.7+8-LTS-224`
in order to benefit from lighter compiler's assisted black-holes.

The perfasm analysis shown the presence of the mentioned calls and the original issue solved here reported already the relevant JDK src code paths involved.

I've decided for 64 bytes as a cutoff value to honor AVX512, but probably even higher values could benefit from this (in short, I've been conservative with the original code). It's interesting to note it by checking the difference between 64  and 65 bytes results.

I didn't reported heap/unpooled data because not relevant, given that is 1:1 in term of perf increase.